Moved binding stage to separate file
